% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{country}
\alias{country}
\title{country}
\usage{
country()
}
\value{
a dataframe with name country and url
}
\description{
a function that retrieves the list of countries available at the url "https://www.monsieur-des-drapeaux.com/drapeaux-du-monde.html"
}
\examples{
world <- country()
head(world)
}
